requirements.txt
numpy==1.18.0
pyyaml
pandas
sklearn
kfp
